refactor(cli): split socket protocol payloads
